more code cleanup, refactor and improvements.
