+++ testSite/DatabaseSearch.php
@@ -232,6 +232,119 @@ class DatabaseSearch {
     }
 
     /**
+     * Echos a data table to show the results for this database search.
+     * @param FileMaker_Result $result
+     */
+    function echoDataTable(FileMaker_Result $result) {
+
+        # filter out unnecessary fields
+        $ignoredFields = ['SortNum', 'Accession Numerical', 'Imaged', 'IIFRNo', 'Photographs::photoFileName',
+            'Event::eventDate', 'card01', 'Has Image', 'imaged'];
+        $usefulFields = array_diff($this->result_layout->listFields(), $ignoredFields);
+
+        # basic table setup with helper functions
+        echo '
+         <!-- construct table for given layout and fields -->
+        <div class="container-fluid no-padding">
+            <!-- id table for special color -->
+            <table class="table table-hover table-striped table-responsive" id="table">
+                <thead>
+                    <tr>
+                        ';
+        $this->echoTableHeads($usefulFields);
+        echo '
+                    </tr>
+                </thead>
+                <tbody>
+                    ';
+        $this->echoTableRows($result->getRecords(), $usefulFields);
+        echo '
+                </tbody>
+            </table>
+        </div>
+    ';
+    }
+    /**
+     * Echos all rows for a table given the records and fields in use.
+     * @param $records
+     * @param $fields
+     */
+    private function echoTableRows($records, $fields) {
+        foreach ($records as $record) {
+            # each record has its table row
+            echo '<tr>';
+            foreach ($fields as $field) {
+                # ID field logic
+                if (formatField($field) === 'Accession Number' or $field === 'SEM #') {
+                    $url = htmlspecialchars($this->name). '&AccessionNo='.htmlspecialchars($record->getField($field));
+                    $id = htmlspecialchars(trim($record->getField($field)));
+
+                    $hasImage = false;
+                    if($this->name === 'entomology' and $record->getField("Imaged") === "Photographed") $hasImage = true;
+                    else if ($this->name === 'fish' and $record->getField("imaged") === "Yes") $hasImage = true;
+                    else if ($this->name === 'mammal' or $this->name === 'avian' or $this->name === 'herpetology') {
+                        if ($record->getField("Photographs::photoFileName") !== "") $hasImage = true;
+                    }
+                    # for vwsp lichen bryophytes fungi algae
+                    else if ($record->getField("Imaged") === "Yes") $hasImage = true;
+
+                    echo "
+                        <th scope='row' id='data'>
+                            <a href='details.php?Database=$url'>
+                            " . ($hasImage ? '<span style="display:inline" id = "icon"  class="oi oi-image"></span>' : '') . "
+                            <b>$id</b>
+                        </th>
+                    ";
+                }
+                # genus or species field for special style
+                else if (formatField($field) == 'Genus' || formatField($field) == 'Species') {
+                    echo '<td id="data" style="font-style:italic;">' . htmlspecialchars($record->getField($field)) . '</td>';
+                }
+                else {
+                    echo '<td id="data">'. $record->getField($field) . '</td>';
+                }
+            }
+            echo '</tr>';
+        }
+    }
+    /**
+     * Echos the table heads given the fields in use.
+     * Will also link the heads with sort urls.
+     * TODO remove all $_GET and $_SERVER calls
+     * @param $fields
+     */
+    private function echoTableHeads($fields) {
+        $page = $_GET['Page'] ?? 1;
+
+        foreach($fields as $field) {
+            $id = htmlspecialchars(formatField($field));
+
+            $payloadList = [
+                'Database' => $this->name,
+                'Sort' => $field,
+                'SortOrder' => $_GET['SortOrder'] ?? 'Descend' == 'Descend' ? 'Ascend' : 'Descend',
+                'Page' => $page,
+            ];
+
+            $href = substr($_SERVER['REQUEST_URI'], 0, strpos($_SERVER['REQUEST_URI'], '?')) . '?' . http_build_query($payloadList);
+            $href = str_replace('%3A', ':', $href);
+
+            $icon_class = $_GET['SortOrder'] ?? '' === 'Descend' ? 'oi-sort-descending' : 'oi-sort-ascending';
+
+            echo "
+                <th scope='col' id=$id>
+                    <a style='padding: 0; white-space:nowrap;' href=$href>
+                        <!-- order icon -->
+                        <span id='icon'  class='oi $icon_class'></span>
+                        <!-- field name -->
+                        <b>$id</b>
+                    </a>
+                </th>
+            ";
+        }
+    }
+
+    /**
      * Returns the ID or accession number field name, it is different for different databases.
      * Optional $isNumeric field will change the name used for some databases.
      * @param bool $isNumeric
