+++ testSite/details.php
@@ -1,7 +1,7 @@
 <?php
 
 require_once ('FileMaker.php');
-require_once ('functions.php');
+require_once('utilities.php');
 require_once ('lib/simple_html_dom.php');
 require_once ('credentials_controller.php');
 require_once ('constants.php');
@@ -224,7 +224,7 @@ $record = $allRecordsFound[0];
                                 }
                                 else if (DATABASE == 'vwsp' || DATABASE == 'bryophytes' || DATABASE == 'fungi'
                                 || DATABASE == 'lichen' || DATABASE == 'algae') {
-                                    $url = getPhotoUrl($_GET['AccessionNo']);
+                                    $url = getPhotoUrl($_GET['AccessionNo'], DATABASE);
                                     $validDb = true;
                                 }
                                 if ($validDb) {
@@ -297,38 +297,39 @@ $record = $allRecordsFound[0];
 
     <?php FooterWidget('images/beatyLogo.png') ;?>
 
-    <script >
-     // js slideshow
-      var slideIndex = 1;
-      showSlides(slideIndex);
+    <script>
+        // js slideshow
+        let slideIndex = 1;
+        showSlides(slideIndex);
 
-      // Next/previous controls
-      function plusSlides(n) {
-        showSlides(slideIndex += n);
-      }
+        // Next/previous controls
+        function plusSlides(n) {
+            showSlides(slideIndex += n);
+        }
 
-      // Thumbnail image controls
-      function currentSlide(n) {
-        showSlides(slideIndex = n);
-      }
+        // Thumbnail image controls
+        function currentSlide(n) {
+            showSlides(slideIndex = n);
+        }
 
-      function showSlides(n) {
-        var i;
-        var slides = document.getElementsByClassName("mySlides");
-        var dots = document.getElementsByClassName("dot");
-        if (n > slides.length) {slideIndex = 1}
-        if (n < 1) {slideIndex = slides.length}
+        function showSlides(n) {
+        let i;
+        const slides = document.getElementsByClassName("mySlides");
+        const dots = document.getElementsByClassName("dot");
+
+        if (n > slides.length) { slideIndex = 1 }
+        if (n < 1) { slideIndex = slides.length }
+
         for (i = 0; i < slides.length; i++) {
             slides[i].style.display = "none";
         }
         for (i = 0; i < dots.length; i++) {
             dots[i].className = dots[i].className.replace(" active", "");
         }
+
         slides[slideIndex-1].style.display = "block";
         dots[slideIndex-1].className += " active";
       }
-
-
     </script>
 
     </body>
